minimal-requirements.txt
# EdgeGuard vLLM 生產部署最小依賴llmcompressor

vllm==0.13.0
transformers==4.56.1
torch==2.9.0
prometheus_client==0.23.1
fastapi==0.127.0
uvicorn==0.40.0
pydantic==2.12.5
jinja2==3.1.6
pyyaml==6.0.3
requests==2.32.5
tqdm==4.67.1
nvidia-cudnn-cu12==9.10.2.21
# pandas/scipy/numpy 由 Dockerfile 指定
sentencepiece==0.2.1
llmcompressor
